requirements.txt
fastapi
uvicorn
python-multipart
Pillow
torch
torchvision
transformers
